requirements.txt
numpy
gmpy2
jax[cpu]
optax
seaborn
matplotlib
-e .
tqdm
onager
optuna==3.1.0
optax
dm-haiku
